Temp Logo Added and files reorganized
