+++ dwfront/src/assets/NavigationBar.jsx
@@ -1,0 +1,34 @@
+import React from 'react'
+import {Link} from 'react-router-dom'
+import './style.css'
+import logo from "./images/logoWeb.png";
+
+const bar = (
+    <header className='navbar'>
+        <nav>
+            <ul>
+                <li className="item0">
+                    <img src={logo} alt="image failed" width="80" height="60"/>
+                </li>
+                <li className="item1">
+                    <Link to="/">Home</Link>
+                </li>
+                <li className="item1">
+                    <Link to="/appointments">Appointments</Link>
+                </li>
+                <li className="item1">
+                    <Link to="/contacts">Contacts</Link>
+                </li>
+                <li className="item1">
+                    <Link to="/about">About</Link>
+                </li>
+            </ul>
+        </nav>
+    </header>
+);
+
+function NavigationBar() {
+    return bar
+}
+
+export default NavigationBar;
