+++ dwfront/src/index.js
@@ -2,9 +2,11 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 
-import LandingPage from "./LandingPage";
+import LandingPage from "./assets/LandingPage";
 
-
+/**
+ Renders react pages;
+ */
 const container = document.getElementById('root');
 const root = createRoot(container);
