+++ dwfront/src/pages/About.jsx
@@ -1,0 +1,23 @@
+import React from "react";
+import '../assets/style.css'
+
+
+const element = (
+    <div>
+        <h1>YOUR DENTIST NAME</h1>
+        <body>
+        <p className='class1'>
+            I hope these large words help you trust this website more
+        </p>
+        <p className='class2'>
+            This is really just a test for fonts
+        </p>
+        </body>
+    </div>
+);
+
+function About() {
+    return element
+}
+
+export default About;
